Authentication service feat: configured a server to consume API credentials and serve the client feat: add ‘server/*’ with functions to create, get, read, verify and serve session feat: server handles authenticated routes, serving the client with the authenticated data feat: no localStorage exposure, ‘getAuthHeaders’ now read the session to get token and key feat: add 'createApiKey' to get fresh key that we 'establishSession' with in 'loginFn'

+++ src/lib/utils/storage.ts
@@ -1,22 +1,24 @@
-export const STORAGE_KEYS = {
-  TOKEN: 'token',
-  API_KEY: 'apiKey',
-  USER_ID: 'profileId',
-};
+// guard against SSR,
+// so we dont try to access localStorage on the server
+const isBrowser = typeof window !== "undefined";
 
 export const saveKey = <T = unknown>(key: string, value: T): void => {
+  if (!isBrowser) return;
   localStorage.setItem(key, JSON.stringify(value));
 };
 
 export const loadKey = <T = unknown>(key: string): T | null => {
+  if (!isBrowser) return null;
   return JSON.parse(localStorage.getItem(key) || "null");
 };
 
 export const logoutFromStorage = () => {
+  if (!isBrowser) return;
   localStorage.clear();
   window.location.href = "/";
 };
 
 export const removeKey = (key: string) => {
+  if (!isBrowser) return;
   localStorage.removeItem(key);
 };
